+++ src/payloads/hello_world.cpp
@@ -6,10 +6,13 @@ int main(int argc, char** argv)
 {
     bmboot::notifyPayloadStarted();
 
-    printf("Hello world from CPU%d! Our base address is %p and the Payload Argument is %ld.\n",
+    auto abi_version = bmboot::getMonitorAbiVersion();
+
+    printf("Hello world from CPU%d! Our base address is %p and the Payload Argument is %ld. ABI version %d.%d.\n",
            bmboot::getCpuIndex(),
            (void*) &_vector_table,
-           bmboot::getPayloadArgument());
+           bmboot::getPayloadArgument(),
+           abi_version.major, abi_version.minor);
 
     for (;;) {}
 }